+++ src/components/Fields/RangeField/Mark/Mark.styled.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
-import { ColorTokens, getColorFromToken } from '../../../foundation/Colors';
-import { shouldNotForwardProps } from '../../../utils/should-not-forward-props';
+import { ColorTokens, getColorFromToken } from '../../../../foundation/Colors';
+import { shouldNotForwardProps } from '../../../../utils/should-not-forward-props';
 
 export const StyledMark = styled(
   'span',
