+++ src/components/Fields/index.ts
@@ -1,4 +1,5 @@
 export * from './CurrencyField';
+export * from './RangeField';
 export * from './SearchField';
 export * from './TextField';
 export * from './TokenField';
